ST-305: Add new interfact to cache connector using implicit reads.
Current calls to the cache have explicit reads and parse to
Map[String, String]. This new interface allows for the json response
to be parsed to a specified type using implicit reads over the
explicit reads. This aims to reduce the API surface of the service
layer which has API endpoints that will do these conversions explicitly.